+++ src/main/java/conspire/relics/BottledYoYo.java
@@ -1,0 +1,106 @@
+package conspire.relics;
+
+import java.util.function.Predicate;
+
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.CardGroup;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.helpers.FontHelper;
+import com.megacrit.cardcrawl.helpers.PowerTip;
+import com.megacrit.cardcrawl.relics.AbstractRelic;
+import com.megacrit.cardcrawl.rooms.AbstractRoom;
+
+import basemod.abstracts.CustomBottleRelic;
+import basemod.abstracts.CustomSaveable;
+import conspire.patches.BottledYoYoField;
+
+public class BottledYoYo extends AbstractConspireRelic implements CustomBottleRelic, CustomSaveable<Integer> {
+    public static final String ID = "conspire:BottledYoYo";
+    private boolean cardSelected = true;
+    public AbstractCard card;
+
+    public BottledYoYo() {
+        super(ID, AbstractRelic.RelicTier.UNCOMMON, AbstractRelic.LandingSound.CLINK);
+    }
+
+    @Override
+    public Predicate<AbstractCard> isOnCard() {
+        return BottledYoYoField.inBottledYoYo::get;
+    }
+
+    @Override
+    public Class<Integer> savedType() {
+        return Integer.class;
+    }
+
+    @Override
+    public Integer onSave() {
+        if (this.card != null) {
+            return AbstractDungeon.player.masterDeck.group.indexOf(this.card);
+        } else {
+            return -1;
+        }
+    }
+
+    @Override
+    public void onLoad(Integer index) {
+        if (index != null && index >= 0 && index < AbstractDungeon.player.masterDeck.group.size()) {
+            this.card = AbstractDungeon.player.masterDeck.group.get(index);
+            BottledYoYoField.inBottledYoYo.set(this.card, true);
+            this.setDescriptionAfterLoading();
+        }
+    }
+
+    @Override
+    public void onEquip() {
+        cardSelected = false;
+        if (AbstractDungeon.isScreenUp) {
+            AbstractDungeon.dynamicBanner.hide();
+            AbstractDungeon.overlayMenu.cancelButton.hide();
+            AbstractDungeon.previousScreen = AbstractDungeon.screen;
+        }
+        AbstractDungeon.getCurrRoom().phase = AbstractRoom.RoomPhase.INCOMPLETE;
+        CardGroup group = CardGroup.getGroupWithoutBottledCards(AbstractDungeon.player.masterDeck);
+        AbstractDungeon.gridSelectScreen.open(group, 1, DESCRIPTIONS[1] + name + ".", false, false, false, false);
+    }
+
+    @Override
+    public void onUnequip() {
+        if (card != null) {
+            AbstractCard cardInDeck = AbstractDungeon.player.masterDeck.getSpecificCard(card);
+            if (cardInDeck != null) {
+                BottledYoYoField.inBottledYoYo.set(cardInDeck, false);
+            }
+        }
+    }
+
+    @Override
+    public void update() {
+        super.update();
+        if (!this.cardSelected && !AbstractDungeon.gridSelectScreen.selectedCards.isEmpty()) {
+            this.cardSelected = true;
+            this.card = AbstractDungeon.gridSelectScreen.selectedCards.get(0);
+            BottledYoYoField.inBottledYoYo.set(this.card, true);
+            AbstractDungeon.getCurrRoom().phase = AbstractRoom.RoomPhase.COMPLETE;
+            AbstractDungeon.gridSelectScreen.selectedCards.clear();
+            this.setDescriptionAfterLoading();
+        }
+    }
+
+    @Override
+    public String getUpdatedDescription() {
+        return DESCRIPTIONS[0];
+    }
+
+    public void setDescriptionAfterLoading() {
+        this.description = DESCRIPTIONS[2] + FontHelper.colorString(this.card.name, "y") + DESCRIPTIONS[3];
+        this.tips.clear();
+        this.tips.add(new PowerTip(this.name, this.description));
+        this.initializeTips();
+    }
+
+    @Override
+    public AbstractRelic makeCopy() {
+        return new BottledYoYo();
+    }
+}